pyproject.toml
[project]
name = "tdev-ai-day"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "google-adk>=1.19.0",
    "greenlet>=3.2.4",
    "litellm>=1.80.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
